Update GreedoLayoutManager.java
In case of up scrolling some of the views are not getting filled and they remain empty. When row heights are large they are not much visible, but if make row heights small (100 px, 200 px) on fast scroll up whole views are getting empty. Take this solution and problem won't exist anymore.